Changing regular int into long int

+++ 0x02-functions_nested_loops/101-natural.c
@@ -9,7 +9,7 @@
 void main(void)
 {
 	int i = 0;
-	int sum = 0;
+	long int sum = 0;
 
 	for (; i < 1024; i++)
 	{
